+++ test/integration/verify/controls/azure_resource_health_availability_statuses.rb
@@ -1,0 +1,6 @@
+control 'azure_resource_health_availability_statuses' do
+  describe azure_resource_health_availability_statuses do
+    it { should exist }
+    its('locations') { should include 'ukwest' }
+  end
+end