+++ CHANGELOG.md
@@ -1,10 +1,40 @@
 # Changelog
-<!-- latest_release 1.52.0 -->
+<!-- latest_release 1.58.0 -->
+## [v1.58.0](https://github.com/inspec/inspec-azure/tree/v1.58.0) (2021-08-06)
+
+#### Merged Pull Requests
+- F/bastion hosts [#414](https://github.com/inspec/inspec-azure/pull/414) ([sa-progress](https://github.com/sa-progress))
+<!-- latest_release -->
+
+## [v1.57.0](https://github.com/inspec/inspec-azure/tree/v1.57.0) (2021-08-06)
+
+#### Merged Pull Requests
+- Account Storage Resource Fix [#423](https://github.com/inspec/inspec-azure/pull/423) ([sa-progress](https://github.com/sa-progress))
+
+## [v1.56.0](https://github.com/inspec/inspec-azure/tree/v1.56.0) (2021-08-06)
+
+#### Merged Pull Requests
+- Support ad directory objects [#406](https://github.com/inspec/inspec-azure/pull/406) ([sathish-progress](https://github.com/sathish-progress))
+
+## [v1.55.0](https://github.com/inspec/inspec-azure/tree/v1.55.0) (2021-08-05)
+
+#### Merged Pull Requests
+- Support resource health availability statuses [#411](https://github.com/inspec/inspec-azure/pull/411) ([sathish-progress](https://github.com/sathish-progress))
+
+## [v1.54.0](https://github.com/inspec/inspec-azure/tree/v1.54.0) (2021-08-05)
+
+#### Merged Pull Requests
+- Support resource health emerging issues [#413](https://github.com/inspec/inspec-azure/pull/413) ([sathish-progress](https://github.com/sathish-progress))
+
+## [v1.53.0](https://github.com/inspec/inspec-azure/tree/v1.53.0) (2021-08-05)
+
+#### Merged Pull Requests
+- Support ad domain services [#404](https://github.com/inspec/inspec-azure/pull/404) ([sathish-progress](https://github.com/sathish-progress))
+
 ## [v1.52.0](https://github.com/inspec/inspec-azure/tree/v1.52.0) (2021-07-29)
 
 #### Merged Pull Requests
 - Support container groups [#402](https://github.com/inspec/inspec-azure/pull/402) ([sathish-progress](https://github.com/sathish-progress))
-<!-- latest_release -->
 
 ## [v1.51.0](https://github.com/inspec/inspec-azure/tree/v1.51.0) (2021-07-23)